commit_3

+++ src/main/groovy/org/morago/model/TranslatorProfile.java
@@ -4,6 +4,7 @@ import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.HashSet;
 import java.util.Set;
@@ -26,7 +27,7 @@ public class TranslatorProfile {
     private boolean online = false;
 
     @Column(name = "hourly_rate")
-    private Double hourlyRate;
+    private BigDecimal hourlyRate;
 
     @Column(name = "created_at")
     private LocalDateTime createdAt;
